auth-service/requirements.txt
fastapi
uvicorn
pymongo
motor
passlib[bcrypt]
python-jose[cryptography]
email-validator
